refactor: move all the game logic into seperate module
