+++ pyproject.toml
@@ -32,3 +32,6 @@ select = ["E", "F", "I"]
 dev = [
     "pytest>=9.0.2",
 ]
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
