+++ src/battleship_pygame_lan/logic/__init__.py
@@ -1,0 +1,5 @@
+from .board import Board
+from .enums import FieldState, ShipType
+from .models import Ship
+
+__all__ = ["Board", "ShipType", "FieldState", "Ship"]
