+++ src/battleship_pygame_lan/logic/board.py
@@ -1,69 +1,9 @@
 import logging
-import uuid
-from enum import Enum
+
+from .enums import FieldState, ShipType
+from .models import Ship, _Field
 
 logger = logging.getLogger(__name__)
-
-
-class ShipType(Enum):
-    """
-    Defines the available ship types and their length. Ship length is also it's health
-    """
-
-    FourMaster = 4
-    ThreeMaster = 3
-    TwoMaster = 2
-    OneMaster = 1
-
-
-class Ship:
-    """
-    Represents a single ship on the board.
-    Every ship has unique ID, type and current health.
-    """
-
-    def __init__(self, ship_type: ShipType) -> None:
-        self.id = str(uuid.uuid4())
-        self.ship_type: ShipType = ship_type
-        self.health = self.ship_type.value
-
-    def hit(self) -> None:
-        """
-        Decrease the ship's health by 1.
-        """
-        self.health -= 1
-
-    def is_sunk(self) -> bool:
-        """
-        Checks if the ship is still alive (if we can even call it that lol).
-
-        Returns:
-            bool: True if health if 0 or less, False if it's alive!
-        """
-        return self.health <= 0
-
-
-class FieldState(Enum):
-    """
-    Represents the current state of a single field on the board.
-    """
-
-    # numbers can be later changed to colors
-    Empty = 1
-    Taken = 2
-    Missed = 3
-    Hit = 4
-
-
-class _Field:
-    """
-    Class Field represents one specific game field on the board.
-    Holds its current state and a reference to a Ship (if it's there)
-    """
-
-    def __init__(self) -> None:
-        self.state: FieldState = FieldState.Empty
-        self.ship: Ship | None = None
 
 
 class Board:
